requirements.txt
numpy>=1.18.1
yappi>=1.2.5
torch>=1.5.0
tornado>=6.0.4
pandas>=1.0.3
parlai>=1.6.0
psycopg2-binary>=2.8.5
pytest>=5.0.0
pyzmq>=19.0.1
tqdm>=4.48.0
hydra-core>=1.0.0
mephisto>=1.0.3
